+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-10 18:16:22
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-04-11 12:09:26
+// Commit: perf(ui): apply RepaintBoundary on animated ui widgets
 
 class uiComponentState {
-  final String commitRef = '292';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-04-10T18:16:22+05:00');
+  final String commitRef = '312';
+  final String lastMessage = 'perf(ui): apply RepaintBoundary on animated ui widgets';
+  final DateTime timestamp = DateTime.parse('2026-04-11T12:09:26+05:00');
 }